requirements.txt
django>=5.0
psycopg2-binary
diffusers
torch
torchvision
torchaudio
Pillow
transformers
accelerate
safetensors
python-dotenv
